+++ rs/nervous_system/clients/Cargo.toml
@@ -16,6 +16,7 @@ ic-base-types = { path = "../../types/base_types" }
 ic-error-types = { path = "../../../packages/ic-error-types" }
 ic-ledger-core = { path = "../../ledger_suite/common/ledger_core" }
 ic-management-canister-types-private = { path = "../../types/management_canister_types" }
+ic-nervous-system-canisters = { path = "../canisters" }
 ic-nervous-system-common = { path = "../../nervous_system/common" }
 ic-nervous-system-proxied-canister-calls-tracker = { path = "../proxied_canister_calls_tracker" }
 ic-nervous-system-runtime = { path = "../runtime" }
