+++ rs/nervous_system/clients/src/ledger_client.rs
@@ -4,7 +4,7 @@ use dfn_candid::{ArgumentDecoder, ArgumentEncoder};
 use dfn_core::CanisterId;
 use ic_base_types::PrincipalId;
 use ic_ledger_core::{block::BlockIndex, Tokens};
-pub use ic_nervous_system_common::ledger::ICRC1Ledger;
+pub use ic_nervous_system_canisters::ledger::ICRC1Ledger;
 use ic_nervous_system_common::NervousSystemError;
 use icrc_ledger_client::{ICRC1Client, Runtime};
 use icrc_ledger_types::icrc1::{
