+++ rs/nervous_system/common/test_utils/src/lib.rs
@@ -10,10 +10,8 @@ use futures::{
     },
     StreamExt,
 };
-use ic_nervous_system_common::{
-    ledger::{ICRC1Ledger, IcpLedger},
-    NervousSystemError,
-};
+use ic_nervous_system_canisters::ledger::{ICRC1Ledger, IcpLedger};
+use ic_nervous_system_common::NervousSystemError;
 use icp_ledger::{AccountIdentifier, Tokens};
 use icrc_ledger_types::icrc1::account::Account;
 use std::sync::{atomic, atomic::Ordering as AtomicOrdering, Arc, Mutex};
