+++ rs/sns/swap/src/types.rs
@@ -17,7 +17,8 @@ use crate::{
 use ic_base_types::{CanisterId, PrincipalId};
 use ic_canister_log::log;
 use ic_ledger_core::Tokens;
-use ic_nervous_system_common::{ledger::ICRC1Ledger, ONE_DAY_SECONDS};
+use ic_nervous_system_canisters::ledger::ICRC1Ledger;
+use ic_nervous_system_common::ONE_DAY_SECONDS;
 use ic_nervous_system_proto::pb::v1::Principals;
 use ic_nervous_system_runtime::DfnRuntime;
 use ic_sns_governance::pb::v1::{ClaimedSwapNeuronStatus, NeuronId};
