+++ rs/sns/swap/tests/common/doubles.rs
@@ -1,7 +1,8 @@
 use async_trait::async_trait;
 use ic_base_types::{CanisterId, PrincipalId};
 use ic_ledger_core::Tokens;
-use ic_nervous_system_common::{ledger::ICRC1Ledger, NervousSystemError};
+use ic_nervous_system_canisters::ledger::ICRC1Ledger;
+use ic_nervous_system_common::NervousSystemError;
 use ic_nervous_system_common_test_utils::SpyLedger;
 use ic_sns_governance::pb::v1::{
     manage_neuron_response, manage_neuron_response::ClaimOrRefreshResponse,
